feat(shops): ajout item to a shop
This commit adds the "kotlin-parcelize" plugin to the build.gradle.kts file. It also adds an activity for sellers in the AndroidManifest.xml file. The StoreApiService interface now has a suspend function called createProduct that takes in additional parameters and returns a Response<Void>. The AdminRepository interface no longer returns a String for the createProduct function, and instead just takes in a Product parameter. The AdminRepositoryImpl class now logs exceptions when calling API functions. The StoreRepositoryImpl class now logs successful creation of products and errors when creating products. The SellerViewModel class no longer takes in unnecessary dependencies. The Product and ProductStore classes are now Parcelable. The Store class is also Parcelable and Serializable. Lastly, the SellerView activity extends BaseActivity instead of AppCompatActivity.