chore(storage): Add userName parameter to saveUserLogin and getUserLogin
This commit adds a new parameter, `userName`, to the `saveUserLogin` function in the `UserPreferences` class. It also modifies the return type of the `getUserLogin` function to include the user's name as a third element in the returned Triple. Additionally, a new function `getUserName` is added to retrieve the user's name from shared preferences.

The changes in this commit allow for storing and retrieving the user's name along with their login information. This is used in MainActivity to display a welcome message that includes the user's name.